add lib/processing/touch.rb
