+++ lib/processing.rb
@@ -12,4 +12,5 @@ require 'processing/window'
 require 'processing/vector'
 require 'processing/image'
 require 'processing/font'
+require 'processing/touch'
 require 'processing/processing'
