+++ lib/processing/processing.rb
@@ -13,30 +13,6 @@ module Processing
     RAD2DEG__ = 180.0 / Math::PI
 
 
-    # Touch object.
-    #
-    class Touch
-
-      # Identifier of each touch
-      #
-      attr_reader :id
-
-      # Horizontal position of touch
-      #
-      attr_reader :x
-
-      # Vertical position of touch
-      #
-      attr_reader :y
-
-      # @private
-      def initialize(id, x, y)
-        @id, @x, @y = id, x, y
-      end
-
-    end# Touch
-
-
     # Shader object.
     #
     class Shader
